lambdas/vietocr-processor/requirements.txt
botocore<1.35.0
boto3==1.34.64
Pillow==10.3.0
pdf2image==1.17.0
numpy==1.26.4
vietocr==0.3.12

